requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.5
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.2
coverage==5.5
crayons==0.4.0
cryptography==3.4.6
defusedxml==0.7.1
distlib==0.3.1
dj-rest-auth==2.1.4
Django==3.1.5
django-allauth==0.44.0
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.2
docformatter==1.4
filelock==3.0.12
flake8==3.8.4
gunicorn==20.1.0
identify==2.2.3
idna==2.10
iniconfig==1.1.1
isort==5.7.0
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.0
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
pre-commit==2.12.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.2.0
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
untokenize==0.1.1
urllib3==1.26.4
virtualenv==20.4.3
webdriver-manager==3.4.0
